Add initial support for building MOX U-boot
Building U-Boot for Turris MOX is quite complicated story. Three pieces
have to be put together to the image:

* ATF (ARM Trusted Firmware)
* WTMI (CM3 image)
* U-Boot

The CM3 image has to be signed with asymetric keys. The current version
counts on the fact that the boards do not enforce signed images and
therefore we can go forward with building unsigned versions ourselves.
Later on we will have to simply download signed image and signature from
CZ.NIC. ATF and U-Boot are not signed and we are free to put there
anything we want.