Omnia: Fix default MAC address
Fix default MAC address generation in boot script.
In OpenWRT/TurrisOS the NICs take the MAC addresses from U-Boot that
sets the adresses in C code from data in I2C EEPROM to the NIC
registers. When Debian kernel starts it shuts down clocks for NICs
which means that the MAC addresses set by U-Boot are lost.

Add FDT modififaction to booscript that generate the MAC address
records to the FDT before the kernel is booted, so the MAC
get passed to the kernel in DT.