Replace static DTBs with the distro DTB
Replace static DTBs that we distributed in /boot from files directory in
this repo with the distro DTB.
Refactor SFP/PHY selection procedure to use fdt editing in U-Boot to
disable PHY and enable SFP when SFP is present on boot.
Refactor install and provide boot.scr script for SATA boot, which is
probably impossible to create on install in Omnia Rescue because the
rescue image does not have chroot busybox applet.